:hourglass: update benches
