+++ benches/bench_f64_return_nan.rs
@@ -1,0 +1,43 @@
+#![feature(stdsimd)]
+
+extern crate dev_utils;
+
+use argminmax::ArgMinMax;
+use codspeed_criterion_compat::*;
+use dev_utils::{config, utils};
+
+#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+use argminmax::{SIMDArgMinMax, AVX2, AVX512, SSE};
+use argminmax::{ScalarArgMinMax, SCALAR};
+
+fn nanargminmax_f64_random_array_long(c: &mut Criterion) {
+    let n = config::ARRAY_LENGTH_LONG;
+    let data: &[f64] = &utils::get_random_array::<f64>(n, f64::MIN, f64::MAX);
+    c.bench_function("scalar_nanargminmax_f64", |b| {
+        b.iter(|| SCALAR::argminmax(black_box(data)))
+    });
+    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+    if is_x86_feature_detected!("sse4.1") {
+        c.bench_function("sse_nanargminmax_f64", |b| {
+            b.iter(|| unsafe { SSE::argminmax(black_box(data)) })
+        });
+    }
+    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+    if is_x86_feature_detected!("avx2") {
+        c.bench_function("avx2_nanargminmax_f64", |b| {
+            b.iter(|| unsafe { AVX2::argminmax(black_box(data)) })
+        });
+    }
+    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+    if is_x86_feature_detected!("avx512f") {
+        c.bench_function("avx512_nanargminmax_f64", |b| {
+            b.iter(|| unsafe { AVX512::argminmax(black_box(data)) })
+        });
+    }
+    c.bench_function("impl_nanargminmax_f64", |b| {
+        b.iter(|| black_box(data.argminmax()))
+    });
+}
+
+criterion_group!(benches, nanargminmax_f64_random_array_long,);
+criterion_main!(benches);
